Removes unnecessary login-form.html
I attempted to split the login form into two views - one with links to the various login options, and the other with a login form for local accounts. The login_required decorator by default redirects to /accounts/login, adding complexity.

I may come back to this, because I don't want to encourage users to type a password here. But at this time, it's just simpler to have the login form show an OAuth link.

That way, invalid login data will show the errors on the form.