Removes DukeDS api_key from DukeDSUser model
Breaks tests that expect a user's permanent api_key, but raises exceptions in cases where this is needed

We will not be relying on user's permanent DukeDS API keys going forward
Instead, we will use api_tokens from Duke DS at `/api/v1/user/api_token`

- For web-interactive authentication, we can bear an OAuth token to this endpoint to obtain an api_token
- For command-line tool authentication, we will accept and validate an existing DukeDS api_token

In both cases, the tokens are rather short-lived, but as the transfer API is implemented we will not have a need for offline DukeDS access. And for times we do, we can get a new token by bearing an OAuth token.