Removes assumed delivery.from_user as the identity for DukeDS API calls.
When a user tried to view/accept a delivery, D4S2 would try to get details of the transfer from DukeDS using the from_user's identity. While this worked out in simple tests with quick turnaround, OAuth tokens expire. This means that by the time the recipient clicked the link to view the transfer, the sender's credential may be expired, causing D4S2 to be unable to look up the delivery.

This change adds a user argument to many objects (DeliveryDetails, Message) that are used on behalf of the logged-in user, expecting a Django user.