Remove `CSRF_COOKIE_HTTPONLY=True` from docker production settings
Current efforts to tighten up authentication require the ember application to read the csrf token from cookie storage, and setting `CSRF_COOKIE_HTTPONLY=True` prevents that.

Per https://docs.djangoproject.com/en/2.0/ref/settings/#csrf-cookie-httponly,

> Designating the CSRF cookie as HttpOnly doesn’t offer any practical protection because CSRF is only to protect against cross-domain attacks. If an attacker can read the cookie via JavaScript, they’re already on the same domain as far as the browser knows, so they can do anything they like anyway. (XSS is a much bigger hole than CSRF.)
> Although the setting offers little practical benefit, it’s sometimes required by security auditors.
> If you enable this and need to send the value of the CSRF token with an AJAX request, your JavaScript must pull the value from a hidden CSRF token form input on the page instead of from the cookie.

Passing the CSRF token as a hidden form input on HTML would be a challenge in this scenario since the django application that provides the cookie is not serving the HTML that the UI would need to read it.